pyproject.toml
[project]
name = "llm-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.80.9",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "typer>=0.20.0",
]
